# DATACMNS-1366 - Performance optimizations.