Various optimizations to improve CPU and memory consumption.


**Baseline**

```
Benchmark                                                                                       Mode  Cnt        Score         Error  Units
TypicalEntityReaderBenchmark.kotlinDataClass                                                   thrpt   10  5899638,348 ±   53829,186  ops/s
TypicalEntityReaderBenchmark.kotlinDataClassWithDefaulting                                     thrpt   10  2690653,960 ±   12988,854  ops/s
TypicalEntityReaderBenchmark.simpleEntityConstructorArgsCreation                               thrpt   10  4741691,355 ±   20735,714  ops/s
TypicalEntityReaderBenchmark.simpleEntityGeneratedFieldAccess                                  thrpt   10  7749539,715 ± 1845318,977  ops/s
TypicalEntityReaderBenchmark.simpleEntityGeneratedPropertyAccess                               thrpt   10  7859653,071 ±  195289,752  ops/s
TypicalEntityReaderBenchmark.simpleEntityReflectiveConstructorArgsCreation                     thrpt   10  4163905,432 ±  107432,148  ops/s
TypicalEntityReaderBenchmark.simpleEntityReflectivePropertyAccess                              thrpt   10  7437671,460 ±   96405,066  ops/s
TypicalEntityReaderBenchmark.simpleEntityReflectivePropertyAccessWithCustomConversionRegistry  thrpt   10  6911496,323 ±   64707,708  ops/s
```

**After optimizations**

```
Benchmark                                                                                       Mode  Cnt        Score        Error  Units
TypicalEntityReaderBenchmark.kotlinDataClass                                                   thrpt   10  6676400,185 ±  45164,682  ops/s
TypicalEntityReaderBenchmark.kotlinDataClassWithDefaulting                                     thrpt   10  2981338,468 ±  36943,754  ops/s
TypicalEntityReaderBenchmark.simpleEntityConstructorArgsCreation                               thrpt   10  4445213,022 ±  73618,043  ops/s
TypicalEntityReaderBenchmark.simpleEntityGeneratedFieldAccess                                  thrpt   10  8814664,064 ±  70241,114  ops/s
TypicalEntityReaderBenchmark.simpleEntityGeneratedPropertyAccess                               thrpt   10  7832454,991 ± 118872,563  ops/s
TypicalEntityReaderBenchmark.simpleEntityReflectiveConstructorArgsCreation                     thrpt   10  4564509,834 ±  38466,637  ops/s
TypicalEntityReaderBenchmark.simpleEntityReflectivePropertyAccess                              thrpt   10  8281840,412 ±  58915,514  ops/s
TypicalEntityReaderBenchmark.simpleEntityReflectivePropertyAccessWithCustomConversionRegistry  thrpt   10  7326248,088 ±  33168,993  ops/s
```

---

Related tickets: [DATACMNS-1366](https://jira.spring.io/browse/DATACMNS-1366), [DATACMNS-1366](https://jira.spring.io/browse/DATAMONGO-2046).